Use one command for delete cert/key

+++ helper/helper/piv.py
@@ -389,13 +389,6 @@ class SlotNode(RpcNode):
         self.certificate = certificate
         self._refresh = refresh
 
-    def _require_version(self, major, minor, micro):
-        try:
-            require_version(self.session.version, (major, minor, micro))
-            return True
-        except NotSupportedError:
-            return False
-
     def get_data(self):
         return dict(
             id=f"{int(self.slot):02x}",
@@ -406,17 +399,20 @@ class SlotNode(RpcNode):
             else None,
         )
 
-    @action(condition=lambda self: self.certificate)
-    def delete_certificate(self, params, event, signal):
-        self.session.delete_certificate(self.slot)
-        self.session.put_object(OBJECT_ID.CHUID, generate_chuid())
-        self._refresh()
-        self.certificate = None
-        return dict()
+    @action(condition=lambda self: self.certificate or self.metadata)
+    def delete(self, params, event, signal):
+        delete_cert = params.pop("delete_cert", False)
+        delete_key = params.pop("delete_key", False)
 
-    @action(condition=lambda self: self.metadata and self._require_version(5, 7, 0))
-    def delete_key(self, params, event, signal):
-        self.session.delete_key(self.slot)
+        if not delete_cert and not delete_key:
+            raise ValueError("Missing delete option")
+
+        if delete_cert:
+            self.session.delete_certificate(self.slot)
+            self.session.put_object(OBJECT_ID.CHUID, generate_chuid())
+            self.certificate = None
+        if delete_key:
+            self.session.delete_key(self.slot)
         self._refresh()
         return dict()
